+++ cookbooks/pbis-nativex/metadata.rb
@@ -1,9 +1,9 @@
-name             'nativex-pbis'
+name             'pbis-nativex'
 maintainer       'NativeX'
 maintainer_email 'adrian.herrera@nativex.com'
 license          'All rights reserved'
-description      'Installs/Configures nativex-pbis'
+description      'Installs/Configures pbis-nativex'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
-depends		 'nativex-ad'
+depends		 'ad-nativex'