+++ cookbooks/pbis-nativex/recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: nativex-pbis
+# Cookbook Name:: pbis-nativex
 # Recipe:: default
 #
 # Copyright 2014, NativeX 
@@ -7,5 +7,5 @@
 # All rights reserved - Do Not Redistribute
 #
 
-include_recipe	"nativex-pbis::installpbis"
-include_recipe  "nativex-pbis::joindomain"
+include_recipe	"pbis-nativex::installpbis"
+include_recipe  "pbis-nativex::joindomain"